[#109] Refactor: HttpSessionUtils 클래스 정의 및 기타
- 로그인 세션과 관련된 기능을 모아둔 HttpSessionUtils 클래스 정의
- User 클래스에서만 사용하는 Gender enum을 중첩 enum 으로 선언
- ApiUserController에서 InvalidParameterException에 대한 에러 응답 객체를 반환하도록 변경
