--- BE/src/main/java/com/codesquad/team1/signup/repository/Gender.java
@@ -1,6 +1,0 @@
-package com.codesquad.team1.signup.repository;
-
-public enum Gender {
-    FEMALE,
-    MALE;
-}
